In trying to convert some of my existing openscad projects over to this module I ran across the mental block of adding interior threads to a pre-cut hollow, which broke my brain a bit.
I find it more natural to think of adding a threaded hole to a solid as a 'tap' or difference() operation, rather than differencing a hollow and then adding interior threads to it.

For this reason I've created the, `tap(...)` module, which creates a positive of the internal threads suitable for differencing to produce a threaded tap hole.